// back-end/src/dtos/http-errors/http-error-code.enum.ts
export enum HttpErrorCode {
    ERR_UNAUTHORIZED = 'ERR_UNAUTHORIZED',
    ERR_UNAUTHORIZED_INVALID_USER = 'ERR_UNAUTHORIZED_INVALID_USER',
    ERR_UNAUTHORIZED_INVALID_PASSWORD = 'ERR_UNAUTHORIZED_INVALID_PASSWORD',
    ERR_AUTH_TOKEN_EXPIRED = 'ERR_AUTH_TOKEN_EXPIRED',
    ERR_VALIDATION = 'ERR_VALIDATION',
    ERR_INTERNAL_SERVER_ERROR = 'ERR_INTERNAL_SERVER_ERROR',
}